+++ PremiumScraps/Config.cs
@@ -18,7 +18,7 @@ namespace PremiumScraps
         public Config(ConfigFile cfg, List<Scrap> scraps)
         {
             cfg.SaveOnConfigSet = false;
-            diceEvents = cfg.Bind("General", "Dice events", true, "Adds some custom dice rolls to Emergency Dice items. Requires 'Emergency Dice Updated' 1.6.5+ to work, or else it will be automatically false.");
+            diceEvents = cfg.Bind("General", "Dice events", true, "Adds some custom dice rolls to Emergency Dice items. Requires 'Emergency Dice Updated' 1.7.4+ to work, or else it will be automatically false.");
             unluckyPlayersStr = cfg.Bind("General", "Unlucky players", "76561198984467725,76561199094139351,76561198198881967", "Comma separated list of players Steam ID that you want them to be unlucky. Bad things will happen to unlucky players, use this config to take a sweet revenge on your friends...");
             foreach (Scrap scrap in scraps)
             {